Fixed Interp2 Error
Interp2 must interpolate from coarse to fine, not fine to coarse,
otherwise it doesn't converge properly.  L2E now converges with 2nd
order for 3 iterations of P and 2 iterations of U and V.  P converges
only with 1st order with the max error algorithm, and it seems like this
is the cause of the instability.  @cvogl