Added dt to stabilize
It worked.  Added the numerical values back in for p and grad p.
Getting 2nd order convergence for u and v and 1st order convergence for
p, tested out to the 20th time step.  Question:  Is dt supposed to be
multiplied by F as well?  It didn't converge that way.  @cvogl